requirements.txt
numpy==1.19.3
SpeechRecognition==3.8.1
SQLAlchemy==1.4.7
nltk==3.5
tensorflow_gpu==1.15.0
Flask==1.1.2
tflearn==0.3.2
Pillow==9.0.0
scikit_learn==1.0.2
tensorflow==2.7.0
